Revert the rename setup.js <-> cli.js
