+++ iob
@@ -1,2 +1,2 @@
 #!/usr/bin/env node
-require('./lib/cli.js');
+require('./lib/setup.js');
