+++ lib/setup.js
